Refactor all public views on its own module
