+++ app.py
@@ -4,26 +4,24 @@ A Flask application for snapcraft.io.
 The web frontend for the snap store.
 """
 
-import bleach
 import datetime
 import flask
 import hashlib
 import humanize
 import modules.cache as cache
 import modules.authentication as authentication
+import modules.public.views as public_views
+import modules.public.api as public_api
 import os
-import pycountry
-import re
 import socket
 from dateutil import parser, relativedelta
 from flask_openid import OpenID
 from flask_wtf.csrf import CSRFProtect
 from functools import wraps
 from json import dumps
-from math import floor
 from modules.macaroon import MacaroonRequest, MacaroonResponse
 from operator import itemgetter
-from urllib.parse import parse_qs, unquote, urlparse, urlunparse
+from urllib.parse import unquote, urlparse, urlunparse
 from werkzeug.contrib.fixers import ProxyFix
 
 
@@ -61,30 +59,6 @@ csrf = CSRFProtect(app)
 
 
 # Request only stable snaps
-SNAP_DETAILS_URL = ''.join([
-    SNAPCRAFT_IO_API,
-    'snaps/details/{snap_name}',
-    '?channel=stable',
-])
-DETAILS_QUERY_HEADERS = {
-    'X-Ubuntu-Series': '16',
-    'X-Ubuntu-Architecture': 'amd64',
-}
-
-SNAP_METRICS_URL = ''.join([
-    SNAPCRAFT_IO_API,
-    'snaps/metrics',
-])
-METRICS_QUERY_HEADERS = {
-    'Content-Type': 'application/json'
-}
-
-FEATURE_SNAPS_URL = ''.join([
-    SNAPCRAFT_IO_API,
-    'snaps/search',
-    '?confinement=strict,classic&q=&section=featured',
-])
-
 SNAP_PUB_METRICS_URL = ''.join([
     DASHBOARD_API,
     'snaps/metrics',
@@ -97,28 +71,6 @@ ACCOUNT_URL = ''.join([
     DASHBOARD_API,
     'account',
 ])
-
-SNAP_SEARCH_URL = ''.join([
-    SEARCH_API,
-    'snaps/search',
-    '?q={snap_name}&page={page}&size={size}',
-    '&confinement=strict,classic',
-])
-SEARCH_QUERY_HEADERS = {
-    'X-Ubuntu-Frameworks': '*',
-    'X-Ubuntu-Architecture': 'amd64',
-    'Accept': 'application/hal+json'
-}
-
-PROMOTED_QUERY_URL = ''.join([
-    SNAPCRAFT_IO_API,
-    'snaps/search',
-    '?promoted=true',
-    '&confinement=strict,classic',
-])
-PROMOTED_QUERY_HEADERS = {
-    'X-Ubuntu-Series': '16'
-}
 
 METADATA_QUERY_URL = ''.join([
     DASHBOARD_API,
@@ -174,25 +126,6 @@ def _transform_api_data(details):
     )
 
     return details
-
-
-def get_snap_details(snap_name):
-    details_response = cache.get(
-        SNAP_DETAILS_URL.format(snap_name=snap_name),
-        headers=DETAILS_QUERY_HEADERS
-    )
-
-    if details_response.status_code >= 400:
-        message = (
-            'Failed to get snap details for {snap_name}'.format(**locals())
-        )
-
-        if details_response.status_code == 404:
-            message = 'Snap not found: {snap_name}'.format(**locals())
-
-        flask.abort(details_response.status_code, message)
-
-    return details_response.json()
 
 
 def build_image_info(image, image_type):
@@ -260,19 +193,6 @@ def get_snap_status(snap_id):
     return status_response.json()
 
 
-def get_searched_snaps(snap_searched, size, page):
-    searched_response = cache.get(
-        SNAP_SEARCH_URL.format(
-            snap_name=snap_searched,
-            size=size,
-            page=page
-        ),
-        headers=SEARCH_QUERY_HEADERS
-    )
-
-    return searched_response.json()
-
-
 def login_required(func):
     @wraps(func)
     def is_user_logged_in(*args, **kwargs):
@@ -287,51 +207,9 @@ def redirect_to_login():
     return flask.redirect('login?next=' + flask.request.url_rule.rule)
 
 
-def normalize_searched_snaps(search_results):
-    return (
-        search_results['_embedded']['clickindex:package']
-        if search_results['_embedded']
-        else []
-    )
-
-
-def get_public_metrics(snap_name, json):
-    metrics_response = cache.get(
-        SNAP_METRICS_URL.format(snap_name=snap_name),
-        headers=METRICS_QUERY_HEADERS,
-        json=json
-    )
-
-    return metrics_response.json()
-
-
-def get_featured_snaps():
-    featured_response = cache.get(
-        FEATURE_SNAPS_URL,
-        headers=SEARCH_QUERY_HEADERS
-    )
-
-    return normalize_searched_snaps(featured_response.json())
-
-
-def get_promoted_snaps():
-    promoted_response = cache.get(
-        PROMOTED_QUERY_URL,
-        headers=PROMOTED_QUERY_HEADERS
-    )
-
-    return normalize_searched_snaps(promoted_response.json())
-
-
 def get_snap_id(snap_name):
-    snap_details = get_snap_details(snap_name)
+    snap_details = public_api.get_snap_details(snap_name)
     return snap_details['snap_id']
-
-
-def calculate_color(thisCountry, maxCountry, maxColor, minColor):
-    countryFactor = float(thisCountry)/maxCountry
-    colorRange = maxColor - minColor
-    return int(colorRange*countryFactor+minColor)
 
 
 def get_publisher_metrics(json):
@@ -346,30 +224,6 @@ def get_publisher_metrics(json):
     )
 
     return metrics_response.json()
-
-
-def calculate_colors(countries, max_users):
-    for country_code in countries:
-        countries[country_code]['color_rgb'] = [
-            calculate_color(
-                countries[country_code]['percentage_of_users'],
-                max_users,
-                8,
-                229
-            ),
-            calculate_color(
-                countries[country_code]['percentage_of_users'],
-                max_users,
-                64,
-                245
-            ),
-            calculate_color(
-                countries[country_code]['percentage_of_users'],
-                max_users,
-                129,
-                223
-            )
-        ]
 
 
 # Error handlers
@@ -502,75 +356,16 @@ def create_redirect():
     return flask.redirect('https://docs.snapcraft.io/build-snaps')
 
 
-def convert_limit_offset_to_size_page(link):
-    url_parsed = urlparse(link)
-    host_url = (
-        "{base_url}"
-        "?q={q}&limit={limit}&offset={offset}"
-    )
-
-    url_queries = parse_qs(url_parsed.query)
-    q = url_queries['q'][0]
-    size = int(url_queries['size'][0])
-    page = int(url_queries['page'][0])
-
-    return host_url.format(
-        base_url=flask.request.base_url,
-        q=q,
-        limit=size,
-        offset=size*(page-1)
-    )
-
-
-def get_pages_details(links):
-
-    links_result = {}
-
-    if('first' in links):
-        links_result['first'] = convert_limit_offset_to_size_page(
-            links['first']['href']
-        )
-
-    if('last' in links):
-        links_result['last'] = convert_limit_offset_to_size_page(
-            links['last']['href']
-        )
-
-    if('next' in links):
-        links_result['next'] = convert_limit_offset_to_size_page(
-            links['next']['href']
-        )
-
-    if('prev' in links):
-        links_result['prev'] = convert_limit_offset_to_size_page(
-            links['prev']['href']
-        )
-
-    if('self' in links):
-        links_result['self'] = convert_limit_offset_to_size_page(
-            links['self']['href']
-        )
-
-    return links_result
-
-
 # Normal views
 # ===
 @app.route('/')
 def homepage():
-    return flask.render_template(
-        'index.html',
-        featured_snaps=get_featured_snaps()
-    )
+    return public_views.homepage()
 
 
 @app.route('/store')
 def store():
-    return flask.render_template(
-        'store.html',
-        featured_snaps=get_featured_snaps(),
-        page_slug='store'
-    )
+    return public_views.store()
 
 
 @app.route('/discover')
@@ -580,188 +375,17 @@ def discover():
 
 @app.route('/snaps')
 def snaps():
-    return flask.render_template(
-        'promoted.html',
-        snaps=get_promoted_snaps()
-    )
+    return public_views.snaps()
 
 
 @app.route('/search')
 def search_snap():
-    snap_searched = flask.request.args.get('q', default='', type=str)
-    if(not snap_searched):
-        return flask.redirect('/store')
-
-    size = flask.request.args.get('limit', default=10, type=int)
-    offset = flask.request.args.get('offset', default=0, type=int)
-
-    page = floor(offset / size) + 1
-
-    searched_results = get_searched_snaps(snap_searched, size, page)
-
-    context = {
-        "query": snap_searched,
-        "snaps": normalize_searched_snaps(searched_results),
-        "links": get_pages_details(searched_results['_links'])
-    }
-
-    return flask.render_template(
-        'search.html',
-        **context
-    )
+    return public_views.search_snap()
 
 
 @app.route('/<snap_name>')
 def snap_details(snap_name):
-    """
-    A view to display the snap details page for specific snaps.
-
-    This queries the snapcraft API (api.snapcraft.io) and passes
-    some of the data through to the snap-details.html template,
-    with appropriate sanitation.
-    """
-
-    today = datetime.datetime.utcnow().date()
-    week_ago = today - relativedelta.relativedelta(weeks=1)
-
-    details = get_snap_details(snap_name)
-
-    metrics_query_json = [
-        {
-            "metric_name": "installed_base_by_country_percent",
-            "snap_id": details['snap_id'],
-            "start": week_ago.strftime('%Y-%m-%d'),
-            "end": today.strftime('%Y-%m-%d')
-        }
-    ]
-
-    metrics_response = get_public_metrics(
-        snap_name,
-        metrics_query_json
-    )
-
-    users_by_country = normalize_metrics(
-        metrics_response[0]['series']
-    )
-    country_data = build_country_info(users_by_country)
-
-    description = details['description'].strip()
-    paragraphs = re.compile(r'[\n\r]{2,}').split(description)
-    formatted_paragraphs = []
-
-    # Sanitise paragraphs
-    def external(attrs, new=False):
-        url_parts = urlparse(attrs[(None, "href")])
-        if url_parts.netloc and url_parts.netloc != 'snapcraft.io':
-            if (None, "class") not in attrs:
-                attrs[(None, "class")] = "p-link--external"
-            elif "p-link--external" not in attrs[(None, "class")]:
-                attrs[(None, "class")] += " p-link--external"
-
-        return attrs
-
-    for paragraph in paragraphs:
-        callbacks = bleach.linkifier.DEFAULT_CALLBACKS
-        callbacks.append(external)
-
-        paragraph = bleach.clean(paragraph, tags=[])
-        paragraph = bleach.linkify(paragraph, callbacks=callbacks)
-
-        formatted_paragraphs.append(paragraph.replace('\n', '<br />'))
-
-    context = {
-        # Data direct from details API
-        'snap_title': details['title'],
-        'package_name': details['package_name'],
-        'icon_url': details['icon_url'],
-        'version': details['version'],
-        'revision': details['revision'],
-        'license': details['license'],
-        'publisher': details['publisher'],
-        'screenshot_urls': details['screenshot_urls'],
-        'prices': details['prices'],
-        'support_url': details.get('support_url'),
-        'summary': details['summary'],
-        'description_paragraphs': formatted_paragraphs,
-
-        # Transformed API data
-        'filesize': humanize.naturalsize(details['binary_filesize']),
-        'last_updated': (
-            humanize.naturaldate(
-                parser.parse(details.get('last_updated'))
-            )
-        ),
-
-        # Data from metrics API
-        'countries': country_data,
-
-        # Context info
-        'is_linux': (
-            'Linux' in flask.request.headers['User-Agent'] and
-            'Android' not in flask.request.headers['User-Agent']
-        )
-    }
-
-    return flask.render_template(
-        'snap-details.html',
-        **context
-    )
-
-
-def normalize_metrics(geodata):
-    users_by_country = {}
-    max_users = 0.0
-    for country_counts in geodata:
-        country_code = country_counts['name']
-        users_by_country[country_code] = {}
-        counts = []
-        for daily_count in country_counts['values']:
-            if daily_count is not None:
-                counts.append(daily_count)
-
-        if len(counts) > 0:
-            users_by_country[country_code]['number_of_users'] = (
-                sum(counts)
-            )
-            users_by_country[country_code]['percentage_of_users'] = (
-                sum(counts) / len(counts)
-            )
-        else:
-            users_by_country[country_code]['number_of_users'] = 0
-            users_by_country[country_code]['percentage_of_users'] = 0
-
-        if max_users < users_by_country[country_code]['percentage_of_users']:
-            max_users = users_by_country[country_code]['percentage_of_users']
-
-    calculate_colors(users_by_country, max_users)
-
-    return users_by_country
-
-
-def build_country_info(users_by_country, display_number_users=False):
-    country_data = {}
-    for country in pycountry.countries:
-        country_info = users_by_country.get(country.alpha_2)
-        number_of_users = 0
-        percentage_of_users = 0
-        color_rgb = [247, 247, 247]
-        if country_info is not None:
-            if display_number_users:
-                number_of_users = country_info['number_of_users'] or 0
-            percentage_of_users = country_info['percentage_of_users'] or 0
-            color_rgb = country_info['color_rgb'] or [247, 247, 247]
-
-        country_data[country.numeric] = {
-            'name': country.name,
-            'code': country.alpha_2,
-            'percentage_of_users': percentage_of_users,
-            'color_rgb': color_rgb
-        }
-
-        if display_number_users:
-            country_data[country.numeric]['number_of_users'] = number_of_users
-
-    return country_data
+    return public_views.snap_details(snap_name)
 
 
 # Publisher views
@@ -786,7 +410,7 @@ def publisher_snap_measure(snap_name):
         type=str
     )
 
-    details = get_snap_details(snap_name)
+    details = public_api.get_snap_details(snap_name)
 
     today = datetime.datetime.utcnow().date()
     end = today - relativedelta.relativedelta(days=1)
@@ -841,11 +465,11 @@ def publisher_snap_measure(snap_name):
         'buckets': active_devices['buckets']
     }
 
-    users_by_country = normalize_metrics(
+    users_by_country = public_views.normalize_metrics(
         metrics_response_json['metrics'][1]['series']
     )
 
-    country_data = build_country_info(
+    country_data = public_views.build_country_info(
         users_by_country,
         True
     )
@@ -883,7 +507,7 @@ def get_market_snap(snap_name):
     snap_id = get_snap_id(snap_name)
     metadata = snap_metadata(snap_id)
     details = _transform_api_data(
-        get_snap_details(snap_name)
+        public_api.get_snap_details(snap_name)
     )
 
     context = {
@@ -956,7 +580,7 @@ def post_market_snap(snap_name):
             'title',
             'summary',
             'description',
-            'contact_url',
+            'contact',
             'keywords',
             'license',
             'price',
@@ -975,16 +599,16 @@ def post_market_snap(snap_name):
 
         if error_list:
             details = _transform_api_data(
-                get_snap_details(snap_name)
+                public_api.get_snap_details(snap_name)
             )
 
             context = {
                 "snap_id": flask.request.form['snap_id'],
                 "snap_name": snap_name,
-                "title": metadata['title'],
-                "summary": metadata['summary'],
-                "description": metadata['description'],
-                "license": metadata['license'],
+                "title": details['title'],
+                "summary": details['summary'],
+                "description": details['description'],
+                "license": details['license'],
                 "details": details,
                 "screenshots": screenshots_response,
                 "error_list": error_list
