+++ templates/publisher/market.html
@@ -156,7 +156,7 @@
             <label>Developer website: </label>
           </div>
           <div class="col-8">
-            <input type="text" name="contact_url" value="{{ contact_url }}" />
+            <input type="text" name="contact" value="{{ contact }}" />
           </div>
         </div>
       </form>
